🚚 Move array arbitraries into /arbitrary
