+++ src/arbitrary/_internals/helpers/MaxLengthFromMinLength.ts
@@ -1,0 +1,7 @@
+/**
+ * Compute `maxLength` based on `minLength`
+ * @internal
+ */
+export function maxLengthFromMinLength(minLength: number): number {
+  return 2 * minLength + 10;
+}
