+++ src/check/arbitrary/LoremArbitrary.ts
@@ -1,4 +1,4 @@
-import { array } from './ArrayArbitrary';
+import { array } from '../../arbitrary/array';
 import { constant } from './ConstantArbitrary';
 import { Arbitrary } from './definition/Arbitrary';
 import { frequency } from './FrequencyArbitrary';
