+++ src/check/arbitrary/StringArbitrary.ts
@@ -1,4 +1,5 @@
-import { array, maxLengthFromMinLength } from './ArrayArbitrary';
+import { array } from '../../arbitrary/array';
+import { maxLengthFromMinLength } from '../../arbitrary/_internals/helpers/MaxLengthFromMinLength';
 import { ascii, base64, char, char16bits, fullUnicode, hexa, unicode } from './CharacterArbitrary';
 import { Arbitrary } from './definition/Arbitrary';
