+++ src/check/arbitrary/TypedArrayArbitrary.ts
@@ -1,4 +1,4 @@
-import { array } from './ArrayArbitrary';
+import { array } from '../../arbitrary/array';
 import { Arbitrary } from './definition/Arbitrary';
 import { DoubleNextConstraints } from './DoubleNextArbitrary';
 import { double, float } from './FloatingPointArbitrary';
