+++ src/fast-check-default.ts
@@ -19,7 +19,7 @@ import {
 import { assert, check } from './check/runner/Runner';
 import { sample, statistics } from './check/runner/Sampler';
 
-import { array, ArrayConstraints } from './check/arbitrary/ArrayArbitrary';
+import { array, ArrayConstraints } from './arbitrary/array';
 import {
   bigInt,
   bigIntN,
@@ -72,7 +72,7 @@ import {
 import { oneof, OneOfValue, OneOfConstraints } from './check/arbitrary/OneOfArbitrary';
 import { option, OptionConstraints } from './check/arbitrary/OptionArbitrary';
 import { record, RecordConstraints, RecordValue } from './check/arbitrary/RecordArbitrary';
-import { set, SetConstraints } from './check/arbitrary/SetArbitrary';
+import { set, SetConstraints } from './arbitrary/set';
 import { infiniteStream } from './check/arbitrary/StreamArbitrary';
 import {
   asciiString,
